% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_gxe_post_process.R
\name{plot.tlGxE}
\alias{plot.tlGxE}
\title{Manhattan Plots for ATE and MOR additive and codominant tests.}
\usage{
\method{plot}{tlGxE}(object, ...)
}
\arguments{
\item{object}{a \code{tlGxE} object, ie, a reference to the \code{tlGxE} function call.}

\item{...}{see details}
}
\description{
Manhattan plots for ATE and MOR additive and codominant tests in \code{tlGxE}.
}
\details{
Pass the desired testing significance level as an argument in the function call.  Default level is \eqn{\alpha = 0.05}, which should
be changed to account for multiple testing.
}
